{{ cookiecutter.project_slug }}/pyproject.toml

[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "{{ cookiecutter.description }}"
authors = ["{{ cookiecutter.author_name }} <seu@email.com>"]
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
python-decouple = "^3.8"
mkdocs = "^1.5"
mkdocs-material = "^9.5"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
django-filter = "^23.3.0"
django-debug-toolbar = "^4.2.0"
django-redis = "^5.4.0"
Pillow = "^10.1.0"
celery = "^5.3.4"
redis = "^5.0.1"
sentry-sdk = "^1.32.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
factory-boy = "^3.3.0"
faker = "^19.13.0"
coverage = "^7.3.2"
django-stubs = "^4.2.6"
model-bakery = "^1.17.0"
pre-commit = "^3.5.0"
bandit = "^1.7.5"
safety = "^2.3.5"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = """
    --doctest-modules
    --cov=.
    --cov-report=html
    --cov-report=term-missing
    -vv
"""
